Add some fixes, according to @visionmedia comments
