+++ lib/response.js
@@ -131,11 +131,9 @@ res.send = function(body){
 
   // ETag support
   // TODO: W/ support
-  if (app.settings['etag']) {
-    if (len > 1024) {
-      if (!this.get('ETag')) {
-        this.set('ETag', etag(body));
-      }
+  if (app.settings['etag'] && len > 1024) {
+    if (!this.get('ETag')) {
+      this.set('ETag', etag(body));
     }
   }
